+++ docs/doc_tools.rb
@@ -121,8 +121,8 @@ REPLACEMENTS = [
   [/("?)'"([a-z_.]+)"'("?)/, '\1\2\3'],
   # hidden parameters to make test work
   [/OPTST_[A-Z0-5]+/, ''],
-  [/LCL_TST_(UTF|ASC)_FILE(NAME|PATH)/, 'testfile.bin'],
-  ['LCL_SYNC_DIR', '/data/localsync'],
+  [/TST_(UTF|ASC)_(FILENAME|LCL_PATH)/, 'testfile.bin'],
+  ['TST_SYNC_LCL_DIR', '/data/localsync'],
   ['HSTS_UPLOADED_FILE', 'testfile.bin'],
   ['HSTS_FOLDER_UPLOAD', 'folder_1'],
   [%q['"CF_LOCAL_SYNC_DIR"'], 'syncdir'],